RavenDB-18554 - Moved IndexCompactionInProgressException and improved tests.
